add oracle trait and update mock
