+++ crates/lib/Cargo.toml
@@ -42,7 +42,9 @@ log = { workspace = true }
 clap = { workspace = true }
 utoipa = { workspace = true }
 dirs = "6.0.0"
+mockall = "0.13.1"
 
 [dev-dependencies]
 tempfile = "3.2"
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
+mockito = "1.2.0"
