+++ crates/lib/src/oracle/jupiter.rs
@@ -1,4 +1,4 @@
-use super::{PriceSource, TokenPrice};
+use super::{PriceOracle, PriceSource, TokenPrice};
 use crate::error::KoraError;
 use reqwest::Client;
 use serde::Deserialize;
@@ -10,6 +10,7 @@ const JUPITER_API_URL: &str = "https://api.jup.ag/price/v2";
 struct JupiterResponse {
     data: HashMap<String, JupiterPriceData>,
     #[serde(rename = "timeTaken")]
+    #[allow(dead_code)]
     time_taken: f64,
 }
 
@@ -21,38 +22,84 @@ struct JupiterPriceData {
     price: String,
 }
 
-pub async fn get_price(client: &Client, mint_address: &str) -> Result<TokenPrice, KoraError> {
-    // Get price in SOL using vsToken parameter
-    let url = format!(
-        "{}/price?ids={}&vsToken=So11111111111111111111111111111111111111112",
-        JUPITER_API_URL, mint_address
-    );
+pub struct JupiterPriceOracle;
 
-    let response = client
-        .get(&url)
-        .send()
-        .await
-        .map_err(|e| KoraError::RpcError(format!("Jupiter API request failed: {}", e)))?;
+#[async_trait::async_trait]
+impl PriceOracle for JupiterPriceOracle {
+    async fn get_price(&self, client: &Client, mint_address: &str) -> Result<TokenPrice, KoraError> {
+        // Get price in SOL using vsToken parameter
+        let url = format!(
+            "{}/price?ids={}&vsToken=So11111111111111111111111111111111111111112",
+            JUPITER_API_URL, mint_address
+        );
 
-    if !response.status().is_success() {
-        return Err(KoraError::RpcError(format!("Jupiter API error: {}", response.status())));
+        let response = client
+            .get(&url)
+            .send()
+            .await
+            .map_err(|e| KoraError::RpcError(format!("Jupiter API request failed: {}", e)))?;
+
+        if !response.status().is_success() {
+            return Err(KoraError::RpcError(format!("Jupiter API error: {}", response.status())));
+        }
+
+        let jupiter_response: JupiterResponse = response
+            .json()
+            .await
+            .map_err(|e| KoraError::RpcError(format!("Failed to parse Jupiter response: {}", e)))?;
+
+        let price_data = jupiter_response
+            .data
+            .get(mint_address)
+            .ok_or_else(|| KoraError::RpcError("No price data from Jupiter".to_string()))?;
+
+        // Convert price from string to f64
+        let price = price_data
+            .price
+            .parse::<f64>()
+            .map_err(|e| KoraError::RpcError(format!("Failed to parse price: {}", e)))?;
+
+        Ok(TokenPrice { price, confidence: 0.95, source: PriceSource::Jupiter })
     }
+}
 
-    let jupiter_response: JupiterResponse = response
-        .json()
-        .await
-        .map_err(|e| KoraError::RpcError(format!("Failed to parse Jupiter response: {}", e)))?;
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use mockito::{Matcher, Server};
 
-    let price_data = jupiter_response
-        .data
-        .get(mint_address)
-        .ok_or_else(|| KoraError::RpcError("No price data from Jupiter".to_string()))?;
+    #[tokio::test]
+    async fn test_jupiter_price_fetch() {
+        let mock_response = r#"{
+            "data": {
+                "So11111111111111111111111111111111111111112": {
+                    "id": "So11111111111111111111111111111111111111112",
+                    "type": "derivedPrice",
+                    "price": "1"
+                },
+                "JUPyiwrYJFskUPiHa7hkeR8VUtAeFoSYbKedZNsDvCN": {
+                    "id": "JUPyiwrYJFskUPiHa7hkeR8VUtAeFoSYbKedZNsDvCN",
+                    "type": "derivedPrice",
+                    "price": "0.005321503266927636"
+                }
+            },
+            "timeTaken": 0.003297425
+        }"#;
+        let mut server = Server::new_async().await;
+        let _m = server.mock("GET", "/price")
+            .match_query(Matcher::Any)
+            .with_status(200)
+            .with_header("content-type", "application/json")
+            .with_body(mock_response)
+            .create();
 
-    // Convert price from string to f64
-    let price = price_data
-        .price
-        .parse::<f64>()
-        .map_err(|e| KoraError::RpcError(format!("Failed to parse price: {}", e)))?;
+        let client = Client::new();
+        let oracle = JupiterPriceOracle;
+        let result = oracle.get_price(&client, "So11111111111111111111111111111111111111112").await;
 
-    Ok(TokenPrice { price, confidence: 0.95, source: PriceSource::Jupiter })
+        assert!(result.is_ok());
+        let price = result.unwrap();
+        assert_eq!(price.price, 1.0);
+        assert_eq!(price.source, PriceSource::Jupiter);
+    }
 }
